requirements.txt
torch
torchvision
numpy
matplotlib
Pillow
scikit-image
lpips